+++ projects/plugins/social/src/js/store/selectors/site-data.js
@@ -1,0 +1,7 @@
+const siteDataSelectors = {
+	getAPIRootUrl: state => state.siteData?.apiRoot ?? null,
+	getAPINonce: state => state.siteData?.apiNonce ?? null,
+	getRegistrationNonce: state => state.siteData?.registrationNonce ?? null,
+};
+
+export default siteDataSelectors;