+++ projects/packages/status/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.13.6] - 2022-05-24
+### Added
+- Allow plugins to filter the list of available modules. Only activate and consider active modules that are available [#24454]
+
+## [1.13.5] - 2022-05-20
+### Changed
+- Modules: Make activate() method Jetpack plugin agnostic. Allowing standalone plugins to use it without Jetpack.
+
 ## [1.13.4] - 2022-05-19
 ### Added
 - PHPCS updates. [#24418]
@@ -184,6 +192,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Packages: Introduce a status package
 
+[1.13.6]: https://github.com/Automattic/jetpack-status/compare/v1.13.5...v1.13.6
+[1.13.5]: https://github.com/Automattic/jetpack-status/compare/v1.13.4...v1.13.5
 [1.13.4]: https://github.com/Automattic/jetpack-status/compare/v1.13.3...v1.13.4
 [1.13.3]: https://github.com/Automattic/jetpack-status/compare/v1.13.2...v1.13.3
 [1.13.2]: https://github.com/Automattic/jetpack-status/compare/v1.13.1...v1.13.2